Switching Azure AD back to B2C.
There is no userprofile API for B2C, we parse the JWT instead.
We can finally switch back to using access token instead of id token,
as was originally intended by the specification.
Disabling MSAL log spam.
Preparing an example of i13n.
Also improved the regex to remove a trailing zero of a non-zero.